requirements.txt
alabaster==0.7.12
astroid==2.3.1
attrs==19.2.0
Babel==2.7.0
backcall==0.1.0
bleach==3.1.0
blis==0.4.1
catalogue==0.2.0
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
Click==7.0
cloudpickle==1.2.2
colorama==0.4.1
cycler==0.10.0
cymem==2.0.3
decorator==4.4.0
defusedxml==0.6.0
dj-database-url==0.5.0
dj-static==0.0.6
docutils==0.15.2
entrypoints==0.3
Flask==1.1.1
future==0.18.2
gunicorn==20.0.4
idna==2.8
imagesize==1.1.0
importlib-metadata==1.3.0
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jedi==0.15.1
jieba==0.40
Jinja2==2.10.3
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.0
keyring==19.2.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.2
line-bot-sdk==1.15.0
MarkupSafe==1.1.1
matplotlib==3.1.1
mccabe==0.6.1
mistune==0.8.4
more-itertools==8.0.2
mpmath==1.1.0
murmurhash==1.0.2
mysql-connector-python==8.0.17
nbconvert==5.6.0
nbformat==4.4.0
nltk==3.4.5
notebook==6.0.1
numpy==1.17.2
numpydoc==0.9.1
opencv-python==4.1.2.30
packaging==19.2
pandas==0.25.2
pandocfilters==1.4.2
parso==0.5.1
pickleshare==0.7.5
Pillow==6.2.1
plac==1.1.3
preshed==3.0.2
prometheus-client==0.7.1
prompt-toolkit==2.0.10
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
Pygments==2.4.2
pylint==2.4.3
PyMuPDF==1.16.10
pyparsing==2.4.2
PyPDF2==1.26.0
PyQt5==5.12.3
PyQt5-sip==12.7.0
PyQtWebEngine==5.12.1
pyrsistent==0.15.4
python-dateutil==2.8.0
pytz==2019.3
pywinpty==0.5.5
pyzmq==18.1.0
QtAwesome==0.6.0
qtconsole==4.5.5
QtPy==1.9.0
requests==2.22.0
rope==0.14.0
scipy==1.3.2
seaborn==0.9.0
Send2Trash==1.5.0
six==1.12.0
snowballstemmer==2.0.0
sounddevice==0.3.14
spacy==2.2.3
SpeechRecognition==3.8.1
Sphinx==2.2.0
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
spycy==0.2
spyder==3.3.6
spyder-kernels==0.5.2
SQLAlchemy==1.3.13
srsly==0.2.0
static3==0.7.0
sympy==1.4
terminado==0.8.2
testpath==0.4.2
thinc==7.3.1
tornado==6.0.3
tqdm==4.41.0
traitlets==4.3.3
typed-ast==1.4.0
urllib3==1.25.6
virtualenv==16.7.6
wasabi==0.5.0
Wave==0.0.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.1
widgetsnbextension==3.5.1
wordcloud==1.6.0
wrapt==1.11.2
zipp==0.6.0
